+++ src/python/module/smartredis/dataset.py
@@ -190,7 +190,7 @@ class Dataset(SRObject):
         """Get the names of all metadata scalars and strings from the DataSet
 
         :return: a list of metadata field names
-        :rtype: list
+        :rtype: list[str]
         """
         return self._data.get_metadata_field_names()
 
@@ -226,6 +226,16 @@ class Dataset(SRObject):
         """Get the names of all tensors in the DataSet
 
         :return: a list of tensor names
-        :rtype: list
+        :rtype: list[str]
         """
         return self._data.get_tensor_names()
+
+    @exception_handler
+    def get_tensor_dims(self, name):
+        """Get the dimensions of a tensor in the DataSet
+
+        :return: a list of the tensor dimensions
+        :rtype: list[int]
+        """
+        typecheck(name, "name", str)
+        return self._data.get_tensor_dims(name)